+++ src/test/java/net/minecraft/class_315.java
@@ -5,6 +5,11 @@ public final class class_315 {
     public final class_7172<Integer> field_1870 = new class_7172<>(16);
     public final class_7172<Integer> field_34959 = new class_7172<>(12);
     public final class_7172<Particle> field_1882 = new class_7172<>(Particle.ALL);
+    public final class_7172<Boolean> field_1888 = new class_7172<>(true);
+    public final class_7172<Integer> field_1878 = new class_7172<>(5);
+    public final class_7172<Cloud> field_1814 = new class_7172<>(Cloud.FANCY);
+    public final class_7172<Double> field_24214 = new class_7172<>(1.0);
 
     public enum Particle { ALL, DECREASED, MINIMAL }
+    public enum Cloud { OFF, FAST, FANCY }
 }